CIRC-2072 Create a facade for instance search
